Move mouse position ray tracing to InputController
